address comments @alvations
