+++ nltk/tokenize/__init__.py
@@ -88,7 +88,7 @@ from nltk.tokenize.toktok import ToktokTokenizer
 from nltk.tokenize.treebank import TreebankWordTokenizer
 from nltk.tokenize.util import string_span_tokenize, regexp_span_tokenize
 from nltk.tokenize.stanford_segmenter import StanfordSegmenter
-from nltk.tokenize.ssp import SonoritySequencingPrinciple
+from nltk.tokenize.sonority_sequencing import SyllableTokenizer
 
 
 # Standard sentence tokenizer.
